+++ addons/unity/EditorPackage/Editor/ChunkyMonkeyWindowRules.cs
@@ -1,4 +1,4 @@
-namespace ChunkyMonkey.Unity
+namespace ChunkyMonkey.GitLfsTools
 {
     internal static class ChunkyMonkeyWindowRules
     {